+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
+- The Policy menu now shows your deployment's name (if logged in)
 - Added new experimental `Semgrep: Secrets` setting to use Secrets product in experimental IDE
 
 ## 1.13.0 - 2025-07-28